convert nParticipants to uint8
This is the length of the participants array.
We have a runtime check that this length is shorter than the max uint8
We can then cast the length of the array to a uint8 internally

+++ packages/nitro-protocol/contracts/CountingApp.sol
@@ -33,7 +33,7 @@ contract CountingApp is IForceMoveApp {
         VariablePart memory a,
         VariablePart memory b,
         uint48, // turnNumB, unused
-        uint256 // nParticipants, unused
+        uint8 // nParticipants, unused
     ) public override pure returns (bool) {
         require(
             appData(b.appData).counter == appData(a.appData).counter + 1,
